1.19.4

+++ common/src/main/java/party/lemons/biomemakeover/entity/adjudicator/phase/FangAttackingPhase.java
@@ -104,7 +104,7 @@ public class FangAttackingPhase extends AttackingPhase
         }
 
         private void conjureFangs(double x, double z, double maxY, double y, float yaw, int warmup) {
-            BlockPos blockPos = new BlockPos(x, y, z);
+            BlockPos blockPos = new BlockPos((int)x, (int)y, (int)z);
             boolean bl = false;
             double d = 0.0D;
