+++ common/src/main/java/party/lemons/biomemakeover/entity/adjudicator/phase/FangBarragePhase.java
@@ -56,7 +56,7 @@ public class FangBarragePhase extends TimedPhase
     }
 
     private void conjureFangs(double x, double z, double maxY, double y, float yaw, int warmup) {
-        BlockPos blockPos = new BlockPos(x, y, z);
+        BlockPos blockPos = new BlockPos((int)x, (int)y, (int)z);
         boolean bl = false;
         double d = 0.0D;
